+++ RELEASE-NOTES.md
@@ -6,6 +6,7 @@ OpenService-BCB
 
 - New request parameter to return `info` field.
 - Fix: document & API output
+- Improvement: update rankings & history in a single batch.
 
 
 0.1.0 - 2017-01-03